+++ crates/consensus/src/lib.rs
@@ -34,7 +34,7 @@ pub mod transaction;
 pub use transaction::BlobTransactionValidationError;
 pub use transaction::{
     SignableTransaction, Transaction, TxEip1559, TxEip2930, TxEip4844, TxEip4844Variant,
-    TxEip4844WithSidecar, TxEnvelope, TxLegacy, TxType, TypedTransaction,
+    TxEip4844WithSidecar, TxEip7702, TxEnvelope, TxLegacy, TxType, TypedTransaction,
 };
 
 pub use alloy_eips::eip4844::{