+++ crates/provider/src/chain.rs
@@ -1,6 +1,6 @@
 use alloy_network::{Ethereum, Network};
 use alloy_primitives::{BlockNumber, U64};
-use alloy_rpc_client::{PollerBuilder, WeakClient};
+use alloy_rpc_client::{NoParams, PollerBuilder, WeakClient};
 use alloy_rpc_types_eth::Block;
 use alloy_transport::{RpcError, Transport};
 use async_stream::stream;
@@ -19,7 +19,7 @@ const NO_BLOCK_NUMBER: BlockNumber = BlockNumber::MAX;
 
 pub(crate) struct ChainStreamPoller<T, N = Ethereum> {
     client: WeakClient<T>,
-    poll_task: PollerBuilder<T, (), U64>,
+    poll_task: PollerBuilder<T, NoParams, U64>,
     next_yield: BlockNumber,
     known_blocks: LruCache<BlockNumber, Block>,
     _phantom: PhantomData<N>,
@@ -39,7 +39,7 @@ impl<T: Transport + Clone, N: Network> ChainStreamPoller<T, N> {
     fn with_next_yield(client: WeakClient<T>, next_yield: BlockNumber) -> Self {
         Self {
             client: client.clone(),
-            poll_task: PollerBuilder::new(client, "eth_blockNumber", ()),
+            poll_task: PollerBuilder::new(client, "eth_blockNumber", []),
             next_yield,
             known_blocks: LruCache::new(BLOCK_CACHE_SIZE),
             _phantom: PhantomData,
@@ -161,7 +161,7 @@ mod tests {
         provider.anvil_mine(Some(U256::from(1)), None).await.unwrap();
 
         let block = with_timeout(stream.next()).await.expect("Block wasn't fetched");
-        assert_eq!(block.header.number, Some(1u64));
+        assert_eq!(block.header.number, 1);
     }
 
     #[tokio::test]