+++ crates/provider/src/ext/trace.rs
@@ -1,6 +1,6 @@
 //! This module extends the Ethereum JSON-RPC provider with the Trace namespace's RPC methods.
 use crate::{Provider, RpcWithBlock};
-use alloy_eips::BlockNumberOrTag;
+use alloy_eips::BlockId;
 use alloy_network::Network;
 use alloy_primitives::TxHash;
 use alloy_rpc_types_eth::Index;
@@ -81,23 +81,20 @@ where
     /// # Note
     ///
     /// Not all nodes support this call.
-    async fn trace_block(
-        &self,
-        block: BlockNumberOrTag,
-    ) -> TransportResult<Vec<LocalizedTransactionTrace>>;
+    async fn trace_block(&self, block: BlockId) -> TransportResult<Vec<LocalizedTransactionTrace>>;
 
     /// Replays a transaction.
     async fn trace_replay_transaction(
         &self,
         hash: TxHash,
-        trace_type: &[TraceType],
+        trace_types: &[TraceType],
     ) -> TransportResult<TraceResults>;
 
     /// Replays all transactions in the given block.
     async fn trace_replay_block_transactions(
         &self,
-        block: BlockNumberOrTag,
-        trace_type: &[TraceType],
+        block: BlockId,
+        trace_types: &[TraceType],
     ) -> TransportResult<Vec<TraceResultsWithTransactionHash>>;
 }
 
@@ -112,10 +109,10 @@ where
     fn trace_call<'a, 'b>(
         &self,
         request: &'a <N as Network>::TransactionRequest,
-        trace_type: &'b [TraceType],
+        trace_types: &'b [TraceType],
     ) -> RpcWithBlock<T, (&'a <N as Network>::TransactionRequest, &'b [TraceType]), TraceResults>
     {
-        RpcWithBlock::new(self.weak_client(), "trace_call", (request, trace_type))
+        RpcWithBlock::new(self.weak_client(), "trace_call", (request, trace_types))
     }
 
     fn trace_call_many<'a>(
@@ -144,9 +141,9 @@ where
     async fn trace_raw_transaction(
         &self,
         data: &[u8],
-        trace_type: &[TraceType],
+        trace_types: &[TraceType],
     ) -> TransportResult<TraceResults> {
-        self.client().request("trace_rawTransaction", (data, trace_type)).await
+        self.client().request("trace_rawTransaction", (data, trace_types)).await
     }
 
     async fn trace_filter(
@@ -156,33 +153,31 @@ where
         self.client().request("trace_filter", (tracer,)).await
     }
 
-    async fn trace_block(
-        &self,
-        block: BlockNumberOrTag,
-    ) -> TransportResult<Vec<LocalizedTransactionTrace>> {
+    async fn trace_block(&self, block: BlockId) -> TransportResult<Vec<LocalizedTransactionTrace>> {
         self.client().request("trace_block", (block,)).await
     }
 
     async fn trace_replay_transaction(
         &self,
         hash: TxHash,
-        trace_type: &[TraceType],
+        trace_types: &[TraceType],
     ) -> TransportResult<TraceResults> {
-        self.client().request("trace_replayTransaction", (hash, trace_type)).await
+        self.client().request("trace_replayTransaction", (hash, trace_types)).await
     }
 
     async fn trace_replay_block_transactions(
         &self,
-        block: BlockNumberOrTag,
-        trace_type: &[TraceType],
+        block: BlockId,
+        trace_types: &[TraceType],
     ) -> TransportResult<Vec<TraceResultsWithTransactionHash>> {
-        self.client().request("trace_replayBlockTransactions", (block, trace_type)).await
+        self.client().request("trace_replayBlockTransactions", (block, trace_types)).await
     }
 }
 
 #[cfg(test)]
 mod test {
     use crate::ProviderBuilder;
+    use alloy_eips::BlockNumberOrTag;
 
     use super::*;
 
@@ -194,7 +189,7 @@ mod test {
     async fn test_trace_block() {
         init_tracing();
         let provider = ProviderBuilder::new().on_anvil();
-        let traces = provider.trace_block(BlockNumberOrTag::Latest).await.unwrap();
+        let traces = provider.trace_block(BlockId::Number(BlockNumberOrTag::Latest)).await.unwrap();
         assert_eq!(traces.len(), 0);
     }
 }