+++ crates/rpc-types-engine/CHANGELOG.md
@@ -5,11 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0](https://github.com/alloy-rs/alloy/releases/tag/v0.2.0) - 2024-07-16
+## [0.3.0](https://github.com/alloy-rs/alloy/releases/tag/v0.3.0) - 2024-08-28
+
+### Bug Fixes
+
+- Remove optimism-related types ([#1203](https://github.com/alloy-rs/alloy/issues/1203))
+
+### Dependencies
+
+- Bump core and rm ssz feat ([#1167](https://github.com/alloy-rs/alloy/issues/1167))
+- Bump jsonrpsee 0.24 ([#1067](https://github.com/alloy-rs/alloy/issues/1067))
+
+### Features
+
+- Add error for pre prague requests ([#1204](https://github.com/alloy-rs/alloy/issues/1204))
+- [engine-types] `PayloadError::PrePragueBlockWithEip7702Transactions` ([#1116](https://github.com/alloy-rs/alloy/issues/1116))
 
 ### Miscellaneous Tasks
 
-- Release 0.2.0
+- Clippy für docs ([#1194](https://github.com/alloy-rs/alloy/issues/1194))
+- [dep] Feature gate jwt in engine types ([#1131](https://github.com/alloy-rs/alloy/issues/1131))
+- Release 0.2.1
 - Release 0.2.0
 - Add payloadbodies v2 to capabilities set ([#1025](https://github.com/alloy-rs/alloy/issues/1025))