+++ crates/transport/CHANGELOG.md
@@ -5,11 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0](https://github.com/alloy-rs/alloy/releases/tag/v0.2.0) - 2024-07-16
+## [0.3.0](https://github.com/alloy-rs/alloy/releases/tag/v0.3.0) - 2024-08-28
+
+### Documentation
+
+- Update links to use docs.rs ([#1066](https://github.com/alloy-rs/alloy/issues/1066))
+
+### Features
+
+- [transport] Retry http errors with 503 status code ([#1164](https://github.com/alloy-rs/alloy/issues/1164))
+- Enable more features transitively in meta crate ([#1097](https://github.com/alloy-rs/alloy/issues/1097))
 
 ### Miscellaneous Tasks
 
-- Release 0.2.0
+- Release 0.2.1
 - Release 0.2.0
 - Fix unnameable types ([#1029](https://github.com/alloy-rs/alloy/issues/1029))